Lab-01/source/requirements.txt
azure-functions
azure-storage==0.36.0
azure-functions-worker==1.0.0b11
singleton-decorator==1.0.0
applicationinsights==0.11.9
numpy==1.18.1
matplotlib==3.1.3
pillow==6.1.0